+++ lib/kitchen/selectors/standard_1.rb
@@ -28,6 +28,7 @@ module Kitchen
         self.unit                       = "div[data-type='unit']"
         self.solution                   = "div[data-type='solution'], " \
                                           "div[data-type='question-solution']"
+        self.injected_question          = "div[data-type='exercise-question']"
       end
 
     end