+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Create `InjectedQuestionElement` and `InjectedQuestionElementEnumerator` classes (minor)
 * Create `BakeInjectedExercise` and `BakeInjectedExerciseQuestion` directions (minor)
 * Update `MoveSolutionsFromExerciseSection` and `MoveSolutionsFromNumberedNote` to move injected solutions (minor)
 * Add `SolutionElementEnumerator` to support the above (minor)